+++ src/main/java/com/vgit/yunqiang/controller/BisOrderVerifyController.java
@@ -115,8 +115,13 @@ public class BisOrderVerifyController {
     @RequestMapping(ControllerConsts.URL_DELETE)
     @ResponseBody
     public Ret delete(Long id) {
-        this.bisOrderService.delete(id);
-        return Ret.me();
+        BisOrder bisOrder = this.bisOrderService.get(id);
+        if (bisOrder.getStatus() == 3) {
+            this.bisOrderService.delete(id);
+            return Ret.me();
+        } else {
+            return Ret.me().setSuccess(false).setInfo("订单不能删除");
+        }
     }
 
 }
